✨ feat: refactor GameCard and PlayPanel components
This commit refactors the `GameCard` and `PlayPanel` components. In the `GameCard` component, the image mapping for card numbers has been updated to start from 0 instead of 1. The old mappings have been replaced with the correct ones.

In the `PlayPanel` component, the rendering of cards has been simplified. Now, only the last card in the `cards` array is displayed using the `GameCard` component. Additionally, the text on the close button has been changed from "Fermer" to "Close".