refactor(playpanel): refactor PlayPanel component
This commit refactors the `PlayPanel` component by removing unnecessary code and comments. The removed code includes a section related to card options that were commented out, as well as some unused code for displaying player information and buttons for next turn. These changes improve the clarity and maintainability of the component.